Add stats calculation for new leg

+++ lib/models/aircraft.dart
@@ -5,9 +5,9 @@ class Aircraft {
   final String id;
   final String registration;
   final String name;
-  final int emptyWeight;
-  final int maxGrossWeight;
-  final int maxPayloadWeight;
+  final double emptyWeight;
+  final double maxGrossWeight;
+  final double maxPayloadWeight;
   final Fuel fuel;
   final double fuelCapacity;
   final int maxSeats;
@@ -34,9 +34,9 @@ class Aircraft {
       id: json['Id'],
       registration: json['Identifier'],
       name: json['AircraftType']['DisplayName'],
-      emptyWeight: json['AircraftType']['emptyWeight'],
-      maxGrossWeight: json['AircraftType']['maximumGrossWeight'],
-      maxPayloadWeight: json['AircraftType']['maximumCargoWeight'],
+      emptyWeight: json['AircraftType']['emptyWeight'].toDouble(),
+      maxGrossWeight: json['AircraftType']['maximumGrossWeight'].toDouble(),
+      maxPayloadWeight: json['AircraftType']['maximumCargoWeight'].toDouble(),
       fuel: FUELS.firstWhere((f) => f.id == json['AircraftType']['fuelType']),
       fuelCapacity: json['AircraftType']['FuelTotalCapacityInGallons'],
       maxSeats: json['AircraftType']['seats'],
